+++ sdk/spring-boot-devtools-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.context.ApplicationContextInitializer=kalix.spring.devtools.impl.KalixContainerInitializer
+org.springframework.context.ApplicationListener=kalix.spring.devtools.impl.KalixContainerInitializer
